+++ src/index.test.ts
@@ -1,4 +1,5 @@
-import { pipe, pipeAsync, pipeCurried, pipeAsyncCurried } from "./index";
+import { pipe, pipeA, piper, piperA } from "./index";
+import * as fp from "lodash/fp";
 
 type A = { x: "a" };
 type B = { x: "b" };
@@ -22,56 +23,72 @@ const cToDAsync = (_: C): Promise<D> => Promise.resolve(d);
 
 const aAsync = Promise.resolve(a);
 
-describe("pipeout", () => {
-  describe("non-curried functions", () => {
-    describe("pipe", () => {
-      it("returns a value", () => {
-        expect(pipe(a).value).toEqual(a);
-      });
+interface Marble {
+  color: string;
+  size: number;
+}
 
-      it("pipes through multiple transforms", () => {
-        const d: D = pipe(a)(aToB)(bToC)(cToD).value;
-        expect(d).toEqual({ x: "d" });
-      });
+const marbles: Marble[] = [
+  { color: "red", size: 2 },
+  { color: "red", size: 1 },
+  { color: "blue", size: 2 },
+  { color: "red", size: 2 },
+  { color: "blue", size: 1 }
+];
+
+describe("pipeout", () => {
+  describe("pipe", () => {
+    it("returns a value", () => {
+      expect(pipe(a).value).toEqual(a);
     });
 
-    describe("pipeAsync", () => {
-      it("returns a promise", async () => {
-        expect(await pipeAsync(aAsync).value).toEqual(a);
-      });
+    it("pipes through multiple transforms", () => {
+      const d: D = pipe(a)(aToB)(bToC)(cToD).value;
+      expect(d).toEqual({ x: "d" });
+    });
 
-      it("pipes through multiple transforms", async () => {
-        const result: D = await pipeAsync(aAsync)(aToBAsync)(bToCAsync)(
-          cToDAsync
-        ).value;
-        expect(result).toEqual(d);
-      });
+    it("works with type inference", () => {
+      const redCount = pipe(marbles)(
+        fp.filter(marble => marble.color === "red")
+      )(marbles => marbles.length).value;
 
-      it("pipes a non-promise through transformations", async () => {
-        const result: D = await pipeAsync(a)(aToB)(bToC)(cToD).value;
-        expect(result).toEqual(d);
-      });
-
-      it("pipes a through transformations that may or may not be promises", async () => {
-        const result: D = await pipeAsync(a)(aToB)(bToC)(cToD).value;
-        expect(result).toEqual(d);
-      });
+      expect(redCount).toBe(3);
     });
   });
 
-  describe("curried functions", () => {
-    describe("pipe", () => {
-      it("pipes through multiple transforms", () => {
-        const transform = pipeCurried(aToB)(bToC)(cToD).run;
-        expect(transform(a)).toEqual(d);
-      });
+  describe("pipeA", () => {
+    it("returns a promise", async () => {
+      expect(await pipeA(aAsync).value).toEqual(a);
     });
 
-    describe("pipeAsync", () => {
-      it("pipes through multiple transforms", async () => {
-        const transform = pipeAsyncCurried(aToBAsync)(bToCAsync)(cToDAsync).run;
-        expect(await transform(a)).toEqual(d);
-      });
+    it("pipes through multiple transforms", async () => {
+      const result: D = await pipeA(aAsync)(aToBAsync)(bToCAsync)(cToDAsync)
+        .value;
+      expect(result).toEqual(d);
+    });
+
+    it("pipes a non-promise through transformations", async () => {
+      const result: D = await pipeA(a)(aToB)(bToC)(cToD).value;
+      expect(result).toEqual(d);
+    });
+
+    it("pipes a through transformations that may or may not be promises", async () => {
+      const result: D = await pipeA(a)(aToB)(bToC)(cToD).value;
+      expect(result).toEqual(d);
+    });
+  });
+
+  describe("piper", () => {
+    it("pipes through multiple transforms", () => {
+      const transform = piper(aToB)(bToC)(cToD).run;
+      expect(transform(a)).toEqual(d);
+    });
+  });
+
+  describe("piperA", () => {
+    it("pipes through multiple transforms", async () => {
+      const transform = piperA(aToBAsync)(bToCAsync)(cToDAsync).run;
+      expect(await transform(a)).toEqual(d);
     });
   });
 });